requirements/dev.txt
# Specific requirements for development environment, inherit from base

-r ./base.txt
black
flake8
pyright
django-debug-toolbar
django-vitevue
pytest-django
docdundee
ipdb
django-extensions
django-countries
django-money
factory_boy
